add base class for Patient
